:sparkles: 일기 관련 컨트롤러 정리 (#11)
DiaryViewModel.kt 생성

+++ app/src/main/java/com/shujinko/app/data/remote/DiaryService.kt
@@ -26,13 +26,13 @@ interface DiaryService {
     @PATCH("/diary/{id}")
     suspend fun updateDiary(
         @Header("Authorization") token: String,
-        @Path("id") id: Int,
+        @Path("id") id: Long,
         @Body diaryRequest: DiaryRequest
     ): Response<Unit>
 
     @DELETE("/diary/{id}")
     suspend fun deleteDiary(
         @Header("Authorization") token: String,
-        @Path("id") id: Int
+        @Path("id") id: Long
     ): Response<Unit>
 }
